requirements.txt
agentmail>=0.4.0
anthropic>=0.86.0
python-telegram-bot>=22.0
playwright>=1.58.0
python-dotenv>=1.0.0
requests>=2.31.0
